modified:   src/plugins/oneko/index.ts 	deleted:    src/plugins/onekoCoins/index.ts 	modified:   src/utils/constants.ts
