+++ src/utils/constants.ts
@@ -403,6 +403,10 @@ export const Devs = /* #__PURE__*/ Object.freeze({
         name: "Grzesiek11",
         id: 368475654662127616n,
     },
+    Gingi: {
+        name: "0xGingi",
+        id: 351859727568994314n,
+    },
 } satisfies Record<string, Dev>);
 
 // iife so #__PURE__ works correctly
